refactor: move tests verifying transition blocking
These are the tests that block the transition and show that we can
continue after the listeners have completed.
